Create instance by group name
Don't bother trying to set the ID when we can set by name.

This previously relied on get-vpc-facts to set a var with the ID. There
was still a reliance on the same name used, but this removes the need
to get the value back from AWS.

Also allows changing the env for the instance created!